feat(web): add MessageInput component
