+++ client-web/src/components/MessageInput/MessageInput.module.scss
@@ -1,0 +1,5 @@
+.form {
+  display: flex;
+  gap: 0.5rem;
+  margin-top: 1rem;
+}
