+++ client-web/src/pages/MessagesPage/MessagesPage.module.scss
@@ -12,11 +12,6 @@
   overflow-y: auto;
 }
 
-.form {
-  display: flex;
-  gap: 0.5rem;
-  margin-top: 1rem;
-}
 
 .header {
   display: flex;
